@@ -1,6 +1,6 @@
 {
   "name": "vuex",
-  "version": "3.1.3",
+  "version": "4.0.0-alpha.1",
   "description": "state management for Vue.js",
   "main": "dist/vuex.common.js",
   "module": "dist/vuex.esm.js",
@@ -20,7 +20,7 @@
     "build:main": "node build/build.main.js",
     "build:logger": "rollup -c build/rollup.logger.config.js",
     "lint": "eslint src test",
-    "test": "npm run lint && npm run test:types && npm run test:unit && npm run test:ssr && npm run test:e2e",
+    "test": "npm run lint && npm run test:unit && npm run test:ssr && npm run test:e2e",
     "test:unit": "rollup -c build/rollup.dev.config.js && jasmine JASMINE_CONFIG_PATH=test/unit/jasmine.json",
     "test:e2e": "node test/e2e/runner.js",
     "test:ssr": "rollup -c build/rollup.dev.config.js && cross-env VUE_ENV=server jasmine JASMINE_CONFIG_PATH=test/unit/jasmine.json",
@@ -40,9 +40,10 @@
   },
   "homepage": "https://github.com/vuejs/vuex#readme",
   "peerDependencies": {
-    "vue": "^2.0.0"
+    "vue": "^3.0.0-alpha.8"
   },
   "devDependencies": {
+    "@vue/compiler-sfc": "3.0.0-alpha.8",
     "babel-core": "^6.22.1",
     "babel-loader": "^7.1.2",
     "babel-plugin-transform-object-rest-spread": "^6.23.0",
@@ -65,8 +66,8 @@
     "terser": "^3.17.0",
     "todomvc-app-css": "^2.1.0",
     "typescript": "^3.7.2",
-    "vue": "^2.5.22",
-    "vue-loader": "^15.2.1",
+    "vue": "3.0.0-alpha.8",
+    "vue-loader": "16.0.0-alpha.3",
     "vue-template-compiler": "^2.5.22",
     "vuepress": "^0.14.1",
     "vuepress-theme-vue": "^1.1.0",
