@@ -1,20 +1,29 @@
 const fs = require('fs')
 const path = require('path')
 const webpack = require('webpack')
-const VueLoaderPlugin = require('vue-loader/lib/plugin')
+const { VueLoaderPlugin } = require('vue-loader')
+
+function buildEntry (dirname) {
+  const lookupDir = path.join(__dirname, dirname)
+
+  return fs.readdirSync(lookupDir).reduce((entries, dir) => {
+    const fullDir = path.join(lookupDir, dir)
+    const entry = path.join(fullDir, 'app.js')
+    if (fs.statSync(fullDir).isDirectory() && fs.existsSync(entry)) {
+      entries[`${dirname}/${dir}`] = ['webpack-hot-middleware/client', entry]
+    }
+
+    return entries
+  }, {})
+}
 
 module.exports = {
   mode: 'development',
 
-  entry: fs.readdirSync(__dirname).reduce((entries, dir) => {
-    const fullDir = path.join(__dirname, dir)
-    const entry = path.join(fullDir, 'app.js')
-    if (fs.statSync(fullDir).isDirectory() && fs.existsSync(entry)) {
-      entries[dir] = ['webpack-hot-middleware/client', entry]
-    }
-
-    return entries
-  }, {}),
+  entry: {
+    ...buildEntry('classic'),
+    ...buildEntry('composition')
+  },
 
   output: {
     path: path.join(__dirname, '__build__'),
