@@ -1,10 +1,12 @@
-import { Store, install } from './store'
+import { createStore, Store } from './store'
+import { useStore } from './injectKey'
 import { mapState, mapMutations, mapGetters, mapActions, createNamespacedHelpers } from './helpers'
 
 export default {
-  Store,
-  install,
   version: '__VERSION__',
+  createStore,
+  Store,
+  useStore,
   mapState,
   mapMutations,
   mapGetters,
@@ -13,8 +15,9 @@ export default {
 }
 
 export {
+  createStore,
   Store,
-  install,
+  useStore,
   mapState,
   mapMutations,
   mapGetters,
