@@ -1,33 +1,18 @@
-export default function (Vue) {
-  const version = Number(Vue.version.split('.')[0])
+import { storeKey } from './injectKey'
 
-  if (version >= 2) {
-    Vue.mixin({ beforeCreate: vuexInit })
-  } else {
-    // override init and inject vuex init procedure
-    // for 1.x backwards compatibility.
-    const _init = Vue.prototype._init
-    Vue.prototype._init = function (options = {}) {
-      options.init = options.init
-        ? [vuexInit].concat(options.init)
-        : vuexInit
-      _init.call(this, options)
+export default function (app, store, injectKey) {
+  app.provide(injectKey || storeKey, store)
+
+  // TODO: Refactor this to use `provide/inject`. It's currently
+  // not possible because Vue 3 doesn't work with `$` prefixed
+  // `provide/inject` at the moment.
+  app.mixin({
+    beforeCreate () {
+      if (!this.parent) {
+        this.$store = typeof store === 'function' ? store() : store
+      } else {
+        this.$store = this.parent.$options.$store
+      }
     }
-  }
-
-  /**
-   * Vuex init hook, injected into each instances init hooks list.
-   */
-
-  function vuexInit () {
-    const options = this.$options
-    // store injection
-    if (options.store) {
-      this.$store = typeof options.store === 'function'
-        ? options.store()
-        : options.store
-    } else if (options.parent && options.parent.$store) {
-      this.$store = options.parent.$store
-    }
-  }
+  })
 }
